src/requirements.txt
aiohttp==3.6.2
aiohttp-jinja2==1.2.0
aliyun-python-sdk-alidns==2.0.13
aliyun-python-sdk-core==2.13.12
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-ecs==4.17.8
aliyun-python-sdk-slb==3.2.16
aliyunsdkcore==1.0.3
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
idna==2.8
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
multidict==4.6.1
PyYAML==5.2
yarl==1.4.2
